Move autoload to separated module.
