+++ RequireJsLoader/config.ts
@@ -223,7 +223,7 @@ define('RequireJsLoader/config', (() => {
                 // Force load polyfills
                 IS_SERVER_SCRIPT ? '' : (wsCoreIncluded ? 'Core/polyfill' : ''),
                 // Force load extra patches for RequireJS
-                'RequireJsLoader/extras'
+                'RequireJsLoader/autoload'
             ], 'RequireJsLoader/_extras/');
 
             let finalDeps = deps;
