+++ RequireJsLoader/extras/autoload.ts
@@ -1,3 +1,3 @@
 // Alias for backward compatibility
-import autoload from '../extras';
+import autoload from '../autoload';
 export default autoload;
